user-story-tests/08-all-page-styles-at-high-priority.html

<!DOCTYPE html>
<html lang="en">

<head>
    <meta charset="UTF-8">
    <meta name="viewport" content="width=device-width, initial-scale=1.0">
    <title>All page styles at high priority</title>
    <script src='./resources/testharness.js'></script>
    <script src='./resources/testharnessreport.js'></script>
</head>

<body>

    <style>
        @layer resets {
            button {
                border: thick dashed red;
            }
        }
    </style>
    <button>Button outside a shadow tree</button>
    <button-group>
        <template shadowrootmode="open">
            <style>
                @layer buttons, inherit;

                @layer buttons {
                    button {
                        border: thick solid black;
                    }
                }
            </style>
            <button>Button inside a shadow tree</button>
        </template>
    </button-group>

    <script type="module">
        import ShadowLayers from '../ShadowLayers.js';

        window.onload = function () {

            const elName = "button-group"

            test(function () {

                const inheritResult = ShadowLayers.inheritFromShadowStatementRule(document.querySelector(elName))
                assert_true(inheritResult, "inherit successfully returns");

                const layerStatementResult = ShadowLayers.findLayerStatementRule(document.querySelector(elName))
                assert_equals(typeof layerStatementResult, "object", "found, so returns object");

                const layerStatementRule = layerStatementResult.styleSheet.cssRules.item(layerStatementResult.ruleIndex)
                assert_array_equals(["buttons", "inherit"],
                    layerStatementRule.nameList,
                    "@layer resets");

            }, "All page styles at high priority");

        }
    </script>

    <div id="log"></div>

</body>

</html>
